requirements.txt
# Django and stuff
Django==1.11.0
django-ajax-selects==1.5.2
django-debug-toolbar==1.7
django-messages-extends==0.5
django-pipeline==1.6.12
libsass==0.12.3

# PostgreSQL connector
psycopg2==2.7.1

# Python bindings to the Pandoc universal document converter
pypandoc==1.3.3

# Common Python requirements
Pygments
PyYAML
pytz
pip

# Production-only minifiers
django-htmlmin==0.10.0
csscompressor==0.9.4
jsmin==2.2.1
